1.修改部分注释 2.增加NetUtils，增加获取ip的方法（支持移动网络和wifi）

+++ utils/src/main/java/com/android/utils/common/StringUtils.java
@@ -4,7 +4,10 @@ import android.text.TextUtils;
 
 /**
  * 字符串工具类
- * Author:Simon
+ *
+ * @author Simon
+ * @version v1.1
+ * @date 2018/2/23
  */
 
 public class StringUtils {
@@ -77,7 +80,7 @@ public class StringUtils {
      */
     public static float safeStrToFloat(String s, float defaultFloat) {
         try {
-            return Integer.parseInt(s);
+            return Float.parseFloat(s);
         } catch (Exception e) {
             return defaultFloat;
         }
@@ -102,7 +105,7 @@ public class StringUtils {
      */
     public static double safeStrToDouble(String s, double defaultDouble) {
         try {
-            return Integer.parseInt(s);
+            return Double.parseDouble(s);
         } catch (Exception e) {
             return defaultDouble;
         }
